# Refactor `Poly` class and its inheritance