--- a/scripts/sparse_poly_performance_test.py
+++ b/scripts/sparse_poly_performance_test.py
@@ -7,49 +7,60 @@ import galois
 
 
 def get_coeffs(degree, N, field):
-    degrees = np.random.randint(0, degree + 1, N, dtype=int)
-    degrees[0] = degree
-    coeffs = field.Random(N, low=1)
-    return degrees, coeffs
+    while True:
+        nonzero_degrees = np.random.randint(0, degree + 1, N, dtype=int)
+        nonzero_degrees[0] = degree
+        nonzero_coeffs = field.Random(N, low=1)
+        if nonzero_degrees.size == np.unique(nonzero_degrees).size:
+            break
+
+    return nonzero_degrees, nonzero_coeffs
 
 
-if True:
-    print("BinaryPoly vs SparsePoly".center(80, "-"))
-    GF = galois.GF2
-    N = 3  # Number of non-zero coefficients
-    degree = 500
-    while degree <= 25_000:
-        print(f"Nonzero: {N} / {degree}, {N/degree*100} %")
-        sp1 = galois.poly.SparsePoly(*get_coeffs(degree, N, GF))
-        sp2 = galois.poly.SparsePoly(*get_coeffs(degree, N, GF))
-        bp1 = galois.poly.BinaryPoly(sp1.integer)
-        bp2 = galois.poly.BinaryPoly(sp2.integer)
-
-        print("  BinaryPoly:\t", end="")
-        ipython.magic("timeit bp1 * bp2")
-
-        print("  SparsePoly:\t", end="")
-        ipython.magic("timeit sp1 * sp2")
-
-        degree *= 2
+GF = galois.GF(2**8)
 
 
-if True:
-    print("DensePoly vs SparsePoly".center(80, "-"))
-    GF = galois.GF(31)
-    N = 3  # Number of non-zero coefficients
-    degree = 500
-    while degree <= 25_000:
-        print(f"Nonzero: {N} / {degree}, {N/degree*100} %")
-        sp1 = galois.poly.SparsePoly(*get_coeffs(degree, N, GF))
-        sp2 = galois.poly.SparsePoly(*get_coeffs(degree, N, GF))
-        dp1 = galois.poly.BinaryPoly(sp1.integer)
-        dp2 = galois.poly.BinaryPoly(sp2.integer)
+# print("DensePoly vs SparsePoly Addition".center(80, "-"))
+# N = 3  # Number of non-zero coefficients
+# degree = 2000
+# while degree <= 32000:
+#     print(f"Nonzero: {N} / {degree}, {N/degree*100} %")
+#     p1 = galois.Poly.Degrees(*get_coeffs(degree, N, GF))
+#     p2 = galois.Poly.Degrees(*get_coeffs(degree, N, GF))
 
-        print("  DensePoly:\t", end="")
-        ipython.magic("timeit dp1 * dp2")
+#     print("  SparsePoly:\t", end="")
+#     p1._type = "sparse"
+#     p2._type = "sparse"
+#     p1.nonzero_degrees, p1.nonzero_coeffs, p2.nonzero_degrees, p2.nonzero_coeffs
+#     ipython.run_line_magic("timeit", "p1 + p2")
 
-        print("  SparsePoly:\t", end="")
-        ipython.magic("timeit sp1 * sp2")
+#     print("  DensePoly:\t", end="")
+#     p1._type = "dense"
+#     p2._type = "dense"
+#     p1.coeffs, p2.coeffs  # Ensure _coeffs is created for arithmetic
+#     ipython.run_line_magic("timeit", "p1 + p2")
 
-        degree *= 2
+#     degree *= 2
+
+
+print("DensePoly vs SparsePoly Multiplication".center(80, "-"))
+N = 3  # Number of non-zero coefficients
+degree = 100
+while degree <= 1000:
+    print(f"Nonzero: {N} / {degree}, {N/degree*100} %")
+    p1 = galois.Poly.Degrees(*get_coeffs(degree, N, GF))
+    p2 = galois.Poly.Degrees(*get_coeffs(degree, N, GF))
+
+    print("  SparsePoly:\t", end="")
+    p1._type = "sparse"
+    p2._type = "sparse"
+    p1.nonzero_degrees, p1.nonzero_coeffs, p2.nonzero_degrees, p2.nonzero_coeffs
+    ipython.run_line_magic("timeit", "p1 * p2")
+
+    print("  DensePoly:\t", end="")
+    p1._type = "dense"
+    p2._type = "dense"
+    p1.coeffs, p2.coeffs  # Ensure _coeffs is created for arithmetic
+    ipython.run_line_magic("timeit", "p1 * p2")
+
+    degree += 100