--- a/tests/polys/test_conversions.py
+++ b/tests/polys/test_conversions.py
@@ -61,10 +61,10 @@ def test_sparse_poly_to_str():
     GF = galois.GF(2**8)
     with GF.display("poly"):
         assert galois._poly_conversion.sparse_poly_to_str([0], GF([0])) == "0"
-        assert galois._poly_conversion.sparse_poly_to_str([3, 1, 0], GF([1, 2, 3])) == "x^3 + (α)x + (α + 1)"
+        assert galois._poly_conversion.sparse_poly_to_str([3, 1, 0], GF([1, 2, 3])) == "x^3 + (α)x + α + 1"
 
         assert galois._poly_conversion.sparse_poly_to_str([0], GF([0]), poly_var="y") == "0"
-        assert galois._poly_conversion.sparse_poly_to_str([3, 1, 0], GF([1, 2, 3]), poly_var="y") == "y^3 + (α)y + (α + 1)"
+        assert galois._poly_conversion.sparse_poly_to_str([3, 1, 0], GF([1, 2, 3]), poly_var="y") == "y^3 + (α)y + α + 1"
 
 
 def test_str_to_sparse_poly():