This PR implements #176. As a consequence of removing `DensePoly`, `BinaryPoly`, and `SparsePoly` and consolidating them into `Poly`, binary array arithmetic/creation is much faster. Below are some examples.

### Polynomial arithmetic over GF(2)

```python
In [1]: import galois

In [2]: f = galois.Poly.Random(20, seed=1); f
Out[2]: Poly(x^20 + x^19 + x^16 + x^15 + x^13 + x^11 + x^9 + x^8 + x^7 + x^5 + x^4 + x^2 + 1, GF(2))

In [3]: g = galois.Poly.Random(10, seed=1); g
Out[3]: Poly(x^10 + x^9 + x^6 + x^5 + x^3 + x, GF(2))

In [4]: f * g
Out[4]: Poly(x^30 + x^28 + x^22 + x^20 + x^19 + x^13 + x^11 + x^10 + x^9 + x^8 + x^5 + x, GF(2))
```

```python
# Before
In [5]: %timeit f * g
196 µs ± 11.4 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

# After
In [5]: %timeit f * g
2.66 µs ± 13 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```

### Polynomial search routines in GF(2)

```python
# Before
In [4]: %timeit galois.primitive_polys(2, 12)
10.1 s ± 123 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

# After
In [2]: %timeit galois.primitive_polys(2, 12)
783 ms ± 1.53 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```